BE History, Wallet, & Profile

+++ app/src/main/java/com/zen/cendakala/data/model/SurveyModel.kt
@@ -22,7 +22,7 @@ data class Question(
 data class AnswerModel(
     val answers: Map<String, Answer>,
     val surveyID: String,
-    val reward: Int
+    val reward: Long
 )
 
 data class Answer(
